wordbookの中身をmainモジュール、coreモジュール、utilsモジュールに分割
